# Conflicts:
#	node/src/chain_spec/mod.rs
#	runtime-modules/proposals/codex/src/lib.rs
#	runtime-modules/proposals/codex/src/proposal_types/parameters.rs
#	runtime-modules/proposals/codex/src/tests/mod.rs
#	runtime/src/tests/proposals_integration/mod.rs
#	runtime/src/tests/proposals_integration/working_group_proposals.rs

+++ runtime-modules/governance/src/mock.rs
@@ -71,7 +71,7 @@ impl election::Trait for Test {
 }
 impl membership::Trait for Test {
     type Event = ();
-    type MemberId = u32;
+    type MemberId = u64;
     type SubscriptionId = u32;
     type PaidTermId = u32;
     type ActorId = u32;
