+++ runtime-modules/membership/src/mock.rs
@@ -80,7 +80,7 @@ impl GovernanceCurrency for Test {
 
 impl Trait for Test {
     type Event = ();
-    type MemberId = u32;
+    type MemberId = u64;
     type PaidTermId = u32;
     type SubscriptionId = u32;
     type ActorId = u32;
